+++ CHANGELOG.md
@@ -16,6 +16,14 @@ The minor version will be incremented upon a breaking change and the patch versi
 
 ### Breaking
 
+## 2025-02-22
+
+- richat-shared-v2.5.0
+
+### Features
+
+- shared: add features ([#77](https://github.com/lamports-dev/richat/pull/77))
+
 ## 2025-02-20
 
 - richat-cli-v2.2.1